delete unused files and create function to upload ppifile
